Bump OSC version to 1.8.0, and document where version appears
Bump the version number of OSC to the upcoming release, 1.8.0, and
document where the version number appears in the sources with a
special `OSC_VERSION` or `OSC_VERSION_BEFORE` tag in comments.

Rationale: When starting a new version, many locations need to be
updated, and searching them by version number is unrelibable.
On the now semi-obsolete `main` branch, `1.5.2` is found 12 times,
including 3 occurences in `go.sum` that should be ignored.
On `devel`, `1.7.0` is found 41 times, and most of them do not
actually refer to the current OSC version number.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>