Enable image signature check for CoCo
The kata-agent should manage the signature verification of container
images pulled in the guest. It should be configured with two options:
- enable_signature_verification=true
- image_policy_file=<url>

While this can be easily achieved with annotations for regular kata
pods, no mechanism exists for peer pods. In the meantime, let's
hardcode the agent config directly in the podvm image for the CoCo
case and have the kata-agent using it.

Fixes: https://issues.redhat.com/browse/KATA-3393

Suggested-by: Wainer dos Santos Moschetta <wainersm@redhat.com>
Signed-off-by: Greg Kurz <groug@kaod.org>