Build and deploy the metrics server
The metrics server is deployed alongside the OSC controller to collect
SLI/SLO oriented metrics. It is currently assumed to be shipped in the
OSC monitor image. This is suboptimal since the OSC repo doesn't build
the OSC monitor image.

Build it and deploy it in the same container image as the controller.
This is clearer and ensures a custom operator build also has a functional
metrics server. This will also avoid the e2e check in the upstream CI to
fail because the OSC monitor image isn't published yet.

Signed-off-by: Greg Kurz <groug@kaod.org>