Allow to override extension name
PR #492 was too optimistic. The heuristic to discriminate OCP and OKD
is fragile and doesn't work in some CI environments.

Reintroduce a way to enforce the name with an environment variable
until we find a more robust way to handle the extension. This is
a temporary workaround. It is not documented on purpose.

It just requires to add `SANDBOXED_CONTAINERS_EXTENSION` to the
`env` section of the controller manager in the CSV as shown below :

                env:
                - name: SANDBOXED_CONTAINERS_EXTENSION
                  value: sandboxed-containers
                - name: PEERPODS_NAMESPACE
                  value: openshift-sandboxed-containers-operator

Signed-off-by: Greg Kurz <groug@kaod.org>