CSV update: use konflux repository for our images
Changing the image references to use the konflux repo.
This is a temporary measure, to validate konflux's nudging process,
where the bundle file gets updated when the images are rebuilt.

We will put back the release registry URL when our release plan
is set on the konflux side. Then konflux should be able to update
the images with the release URL.

Signed-off-by: Julien Ropé <jrope@redhat.com>