azure: blob as source update
Per Microsoft's announcement, starting March 15, 2025, additional
permission is required when using blobs as the source for Azure Compute
Gallery (ACG) images. This updates our CCO request to include the
required permission.

My understanding is that the Azure job that creates the image is
building the podvm based on blob (VHD), so we need to update only the
CCO request.

Fixes: rhjira#KATA-3579

Signed-off-by: Beraldo Leal <bleal@redhat.com>