clean up previous commit
Previous commit removed the only reason for the 'isKataMcpUpdating'
variable to exist.  'isMcoUpdating's value was previously acquired in two
stages and 'isKataMcpUpdating' was a name for one of them.  Now there's
just a single stage left so we remove the variable.

'isInstallationInProgress' is only used at a single place as well but we
keep the variable since it gives an explicit name to the actual meaning
of the 'isMcpUpdating()' check which could otherwise seem non-obvious.  We
just move the variable close to its only place of use.

Signed-off-by: Pavel Mores <pmores@redhat.com>