konflux: avoid rebuilding images every time something is changed
When a build occurs and one of the images of our operator is rebuilt,
the bundle is updated with a PR from Konflux with the nudge mechanism.
Merging this PR must NOT trigger rebuilds, otherwise we end up in a loop
of build/nudges.
This situation happens with the podvm-builder image.

This commit adds a filter on the "on-push" pipeline to rebuild the
podvm-builder image only when some of its own files were modified.

I'm also adding the same rule on the must-gather image, even if it is not
part of the operator's CSV file and doesn't cause such a loop - it just
doesn't make sense to rebuild it eveytime something changes in the repo.

Signed-off-by: Julien Ropé <jrope@redhat.com>