konflux: fix build triggers for the operator
Our operator gets rebuilt when some unrelated files are modified.
This is typically the case when the bundle is updated, which is a problem
as if the operator gets rebuilt when the bundle in updated, this triggers
a loop of build/updates.

This commit modifies the way we check for the files that should be ignored
when deciding when to build. Hopefully, this will fix the problem we are seeing.

Signed-off-by: Julien Ropé <jrope@redhat.com>