konflux: operator build triggers
Moving the check on the config/manager/.* files as part of the "bundle" checks.

When we get a PR for updating the bundle, both the CSV and manager.yaml gets modified.
The first filter (on bundle/.*) will return true because it doesn't match the config/manager/manager.yaml file.
The last filter (on config/manager/.*) will also return true, because it doesn't match the CSV file modification (under bundle/.*).
The result is that the operator build is triggered.

If I want the operator to skip building when the bundle is updated, I need both filters to be in the same "matches" command.

Signed-off-by: Julien Ropé <jrope@redhat.com>