fbc: update render script
The render.sh script was failing to get the image references for images that
were not already released. This commit fixes this by making sure the build
registry (quay.io) is used while 'opm' is executed.
The release registry (registry.redhat.io) is restored after execution, so that
the resulting catalog can be used for an actual release.

Note that if the resulting catalog references unreleased images, those images
will not be accessible to the cluster unless you set some mirroring from
registry.redhat.io to quay.io.

Signed-off-by: Camilla Conte <cconte@redhat.com>
Cherry-picked-by: Julien Ropé <jrope@redhat.com>