podvm-builder: use submodule to get the cloud-api-adaptor sources
We can't run "git clone" in network isolation, so for our downstream
builds, we have to get the sources differently.

This commit uses a submodule to make the cloud-api-adaptor sources available
during build, using a simple "COPY" command in the Dockerfile to get them
into the image.

This submodule conflicts with "controller-gen" execution during the build.
controller-gen uses the go notation "./..." to find packages. But since our
operator has its sources at the root of the repo, controller-gen will look up
dependencies starting from the root, and find conflicting information from the
submodule.

To avoid that behavior, this commit moves the 'main.go' file under cmd/manager,
and changes controller-gen parameters to give it explicit paths for the packages
it should look at, avoiding the use of ./... and preventing the conflict.

Signed-off-by: Julien Ropé <jrope@redhat.com>

operator: mv the "main.go" file to the cmd/ folder

Signed-off-by: Julien Ropé <jrope@redhat.com>