make-test: fix the test script
To create an integration test in Konflux, we need to put the test pipeline
.yaml file in the repository, and then create the test on the Konflux side,
referencing that file.
Because of that, I could not test the script in the real environment, but
only locally.

This commit fixes the different issues that raised up with the first
execution of this pipeline in Konflux:
- install required dependencies
- separate clone and checkout commands as "--revision" doesn't work in this
  environment
- use --argjson and no quotes for numeric types in the output, so that they're
  not considered as strings

Signed-off-by: Julien Ropé <jrope@redhat.com>